+++ tests/test_api.py
@@ -4,7 +4,27 @@ from coloraide import Color, NaN
 import math
 
 
-class TestAPI(unittest.TestCase):
+class Asserts:
+    """Asserts."""
+
+    def assertColorEqual(self, color1, color2, fit=False, precision=5):
+        """Compare two colors."""
+
+        self.assertEqual(
+            color1.to_string(fit=fit, precision=precision),
+            color2.to_string(fit=fit, precision=precision)
+        )
+
+    def assertColorNotEqual(self, color1, color2, fit=False, precision=5):
+        """Compare two colors."""
+
+        self.assertNotEqual(
+            color1.to_string(fit=fit, precision=precision),
+            color2.to_string(fit=fit, precision=precision)
+        )
+
+
+class TestAPI(Asserts, unittest.TestCase):
     """Test API."""
 
     def test_bad_input(self):
@@ -98,6 +118,12 @@ class TestAPI(unittest.TestCase):
         c3 = Color('purple').convert('hsl')
         self.assertEqual(c2, c3.update(c1))
 
+    def test_update_self(self):
+        """Update with self."""
+
+        c1 = Color('orange')
+        self.assertEqual(c1.update(c1), c1)
+
     def test_mutate(self):
         """Test mutate."""
 
@@ -106,6 +132,35 @@ class TestAPI(unittest.TestCase):
         self.assertNotEqual(c1, c2)
         c2.mutate(c1)
         self.assertEqual(c1, c2)
+
+    def test_convert(self):
+        """Test convert."""
+
+        c1 = Color('orange')
+        c2 = c1.convert('hsl')
+        self.assertColorEqual(c2, Color('hsl(39, 100%, 50%)'), precision=0)
+
+    def test_convert_fit(self):
+        """Test convert fit."""
+
+        c1 = Color('color(srgb 2 -1 0)')
+        self.assertFalse(c1.in_gamut())
+        c2 = c1.convert("hsl", fit=True)
+        self.assertTrue(c2.in_gamut())
+        c3 = c1.convert('hsl').fit()
+        self.assertColorEqual(c2, c3)
+
+    def test_convert_fit_clip(self):
+        """Test convert fit."""
+
+        c1 = Color('color(srgb 2 -1 0)')
+        self.assertFalse(c1.in_gamut())
+        c2 = c1.convert("hsl", fit="clip")
+        self.assertTrue(c2.in_gamut())
+        c3 = c1.convert('hsl').fit(method="clip")
+        self.assertColorEqual(c2, c3)
+        c4 = c1.convert("hsl", fit=True)
+        self.assertColorNotEqual(c2, c4)
 
     def test_luminance(self):
         """Test luminance."""
@@ -140,6 +195,14 @@ class TestAPI(unittest.TestCase):
         c1 = Color('orange')
 
         with self.assertRaises(ValueError):
+            c1.get("bad")
+
+    def test_get_bad_chain(self):
+        """Test bad get."""
+
+        c1 = Color('orange')
+
+        with self.assertRaises(ValueError):
             c1.get("hsl.hue.wrong")
 
     def test_set(self):
@@ -169,7 +232,23 @@ class TestAPI(unittest.TestCase):
         c1 = Color('orange')
 
         with self.assertRaises(ValueError):
+            c1.set("bad", 0.5)
+
+    def test_set_bad_chain(self):
+        """Test bad set."""
+
+        c1 = Color('orange')
+
+        with self.assertRaises(ValueError):
             c1.set("hsl.hue.wrong", 0.5)
+
+    def test_set_bad_input(self):
+        """Test bad set."""
+
+        c1 = Color('orange')
+
+        with self.assertRaises(TypeError):
+            c1.set("red", "bad")
 
     def test_overlay(self):
         """Test overlay logic."""
@@ -179,6 +258,23 @@ class TestAPI(unittest.TestCase):
         c3 = c1.overlay(c2)
         self.assertTrue(c1 is not c3)
         self.assertEqual(c1.overlay(c2), Color('color(srgb 0.5 0.5 0.5)'))
+
+    def test_overlay_bad_space(self):
+        """Test overlay logic."""
+
+        c1 = Color('blue').set('alpha', 0.5)
+        c2 = Color('yellow')
+        with self.assertRaises(ValueError):
+            c1.overlay(c2, space="bad")
+
+    def test_overlay_no_alpha(self):
+        """Test overlay logic when color has no alpha."""
+
+        c1 = Color('blue')
+        c2 = Color('yellow')
+        c3 = c1.overlay(c2)
+        self.assertTrue(c1 is not c3)
+        self.assertEqual(c1.overlay(c2), c1)
 
     def test_overlay_nan_alpha(self):
         """Test overlay logic with alpha as `NaN`."""
@@ -243,6 +339,25 @@ class TestAPI(unittest.TestCase):
         color2 = color.fit()
         self.assertIsNot(color, color2)
         self.assertTrue(color2.in_gamut())
+
+    def test_fit_clip(self):
+        """Test fit."""
+
+        color = Color('color(srgb 2 0.5 0.5)')
+        self.assertFalse(color.in_gamut())
+        color2 = color.fit(method="clip")
+        self.assertIsNot(color, color2)
+        self.assertTrue(color2.in_gamut())
+        color3 = color.fit()
+        self.assertColorNotEqual(color2, color3)
+
+    def test_bad_fit(self):
+        """Test fit."""
+
+        color = Color('color(srgb 2 0.5 0.5)')
+        self.assertFalse(color.in_gamut())
+        with self.assertRaises(ValueError):
+            color.fit(method="bad")
 
     def test_fit_in_place(self):
         """Test fit in place."""
@@ -380,3 +495,469 @@ class TestAPI(unittest.TestCase):
 
         self.assertIsNotNone(Color.match('lab(100% 0 0)'))
         self.assertIsNone(Color.match('lab(100% 0 0)', filters=['srgb']))
+
+    def test_mix(self):
+        """Test interpolation."""
+
+        self.assertColorEqual(Color('red').mix('blue', 1), Color("srgb", [0, 0, 1]))
+        self.assertColorEqual(Color('red').mix('blue', 0.75), Color("srgb", [0.25, 0, 0.75]))
+        self.assertColorEqual(Color('red').mix('blue'), Color("srgb", [0.5, 0, 0.5]))
+        self.assertColorEqual(Color('red').mix('blue', 0.25), Color("srgb", [0.75, 0, 0.25]))
+        self.assertColorEqual(Color('red').mix('blue', 0.0), Color("srgb", [1, 0, 0]))
+
+    def test_mix_space(self):
+        """Test color mix in different space."""
+
+        self.assertColorEqual(Color('red').mix('blue', 1, space='lab'), Color("rgb(-0.00065 0.00037 255)"))
+        self.assertColorEqual(Color('red').mix('blue', 0.75, space='lab'), Color("rgb(144.86 -24.872 194.36)"))
+        self.assertColorEqual(Color('red').mix('blue', space='lab'), Color("rgb(193 -29.51 136.17)"))
+        self.assertColorEqual(Color('red').mix('blue', 0.25, space='lab'), Color("rgb(226.9 -24.31 79.189)"))
+        self.assertColorEqual(Color('red').mix('blue', 0.0, space='lab'), Color("rgb(255 0.00015 0.00002)"))
+
+    def test_mix_out_space(self):
+        """Test interpolation."""
+
+        self.assertColorEqual(
+            Color('red').mix('blue', 1, space="lab", out_space="lab"),
+            Color("lab(29.571% 68.304 -112.04)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0.75, space="lab", out_space="lab"),
+            Color("lab(35.75% 71.43 -66.558)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', space="lab", out_space="lab"),
+            Color("lab(41.93% 74.556 -21.079)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0.25, space="lab", out_space="lab"),
+            Color("lab(48.109% 77.682 24.401)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0.0, space="lab", out_space="lab"),
+            Color("lab(54.288% 80.808 69.881)")
+        )
+
+    def test_mix_alpha(self):
+        """Test mixing alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0 / 0.75)').mix('color(srgb 0 0 1 / 0.25)'),
+            Color('rgb(127.5 0 127.5 / 0.5)')
+        )
+
+    def test_mix_premultiplied_alpha(self):
+        """Test premultiplied alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0 / 0.75)').mix('color(srgb 0 0 1 / 0.25)', premultiplied=True),
+            Color('rgb(191.25 0 63.75 / 0.5)')
+        )
+
+    def test_mix_premultiplied_no_alpha(self):
+        """Test premultiplied alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0)').mix('color(srgb 0 0 1)', premultiplied=True),
+            Color('color(srgb 1 0 0)').mix('color(srgb 0 0 1)')
+        )
+
+    def test_mix_in_place(self):
+        """Test mix in place."""
+
+        color = Color('red')
+        color2 = color.mix('blue')
+        self.assertIsNot(color, color2)
+        self.assertColorEqual(color2, Color("srgb", [0.5, 0, 0.5]))
+        color = Color('red')
+        color2 = color.mix('blue', in_place=True)
+        self.assertIs(color, color2)
+        self.assertColorEqual(color, Color("srgb", [0.5, 0, 0.5]))
+
+    def test_mix_nan(self):
+        """Test mixing with NaN."""
+
+        c1 = Color("srgb", [NaN, 1, 1])
+        c2 = Color("srgb", [0.75, 0, 0])
+        self.assertColorEqual(c1.mix(c2), Color("srgb", [0.75, 0.5, 0.5]))
+        c1 = Color("srgb", [0.25, 1, 1])
+        c2 = Color("srgb", [NaN, 0, 0])
+        self.assertColorEqual(c1.mix(c2), Color("srgb", [0.25, 0.5, 0.5]))
+        c1 = Color("srgb", [NaN, 1, 1])
+        c2 = Color("srgb", [NaN, 0, 0])
+        self.assertColorEqual(c1.mix(c2), Color("srgb", [0, 0.5, 0.5]))
+
+    def test_mix_adjust(self):
+        """Test mix adjust method."""
+
+        c1 = Color("color(srgb 0.25 1 1)")
+        c2 = Color("color(srgb 0.75 0 0)")
+        self.assertColorEqual(c1.mix(c2, adjust=["green", "blue"]), Color("srgb", [0.25, 0.5, 0.5]))
+
+    def test_mix_hue_adjust(self):
+        """Test hue adjusting."""
+
+        c1 = Color('rebeccapurple')
+        c2 = Color('lch(85% 100 805)')
+        self.assertColorEqual(
+            c1.mix(c2, 0.25, adjust=["hue"], hue="shorter", space="lch"),
+            Color("rgb(146.74 -3.9476 106.4)")
+        )
+        self.assertColorEqual(
+            c1.mix(c2, 0.25, adjust=["hue"], hue="longer", space="lch"),
+            Color("rgb(-86.84 87.632 170)")
+        )
+        self.assertColorEqual(
+            c1.mix(c2, 0.25, adjust=["hue"], hue="increasing", space="lch"),
+            Color("rgb(146.74 -3.9476 106.4)")
+        )
+        self.assertColorEqual(
+            c1.mix(c2, 0.25, adjust=["hue"], hue="decreasing", space="lch"),
+            Color("rgb(-86.84 87.632 170)")
+        )
+        self.assertColorEqual(
+            c1.mix(c2, 0.25, adjust=["hue"], hue="specified", space="lch"),
+            Color("rgb(112.84 63.966 -28.832)")
+        )
+
+    def test_mix_hue_adjust_bad(self):
+        """Test hue adjusting."""
+
+        c1 = Color('rebeccapurple')
+        c2 = Color('lch(85% 100 805)')
+        with self.assertRaises(ValueError):
+            c1.mix(c2, 0.25, adjust=["hue"], hue="bad", space="lch")
+
+    def test_mix_progress(self):
+        """Test custom progress."""
+
+        progress = lambda x: x * 3  # noqa: E731
+        self.assertColorEqual(
+            Color('red').mix('blue', 1, out_space="lab", space="lab", progress=progress),
+            Color("lab(-19.862% 43.296 -475.88)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0.75, out_space="lab", space="lab", progress=progress),
+            Color("lab(-1.3247% 52.674 -339.44)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0.5, out_space="lab", space="lab", progress=progress),
+            Color("lab(17.213% 62.052 -203)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0.25, out_space="lab", space="lab", progress=progress),
+            Color("lab(35.75% 71.43 -66.558)")
+        )
+        self.assertColorEqual(
+            Color('red').mix('blue', 0, out_space="lab", space="lab", progress=progress),
+            Color("lab(54.288% 80.808 69.881)")
+        )
+
+    def test_interpolate(self):
+        """Test interpolation."""
+
+        self.assertColorEqual(Color('red').interpolate('blue', space="srgb")(1), Color("srgb", [0, 0, 1]))
+        self.assertColorEqual(Color('red').interpolate('blue', space="srgb")(0.75), Color("srgb", [0.25, 0, 0.75]))
+        self.assertColorEqual(Color('red').interpolate('blue', space="srgb")(0.5), Color("srgb", [0.5, 0, 0.5]))
+        self.assertColorEqual(Color('red').interpolate('blue', space="srgb")(0.25), Color("srgb", [0.75, 0, 0.25]))
+        self.assertColorEqual(Color('red').interpolate('blue', space="srgb")(0), Color("srgb", [1, 0, 0]))
+
+    def test_interpolate_space(self):
+        """Test color mix in different space."""
+
+        self.assertColorEqual(Color('red').interpolate('blue', space='lab')(1), Color("rgb(-0.00065 0.00037 255)"))
+        self.assertColorEqual(Color('red').interpolate('blue', space='lab')(0.75), Color("rgb(144.86 -24.872 194.36)"))
+        self.assertColorEqual(Color('red').interpolate('blue', space='lab')(0.5), Color("rgb(193 -29.51 136.17)"))
+        self.assertColorEqual(Color('red').interpolate('blue', space='lab')(0.25), Color("rgb(226.9 -24.31 79.189)"))
+        self.assertColorEqual(Color('red').interpolate('blue', space='lab')(0), Color("rgb(255 0.00015 0.00002)"))
+
+    def test_interpolate_out_space(self):
+        """Test interpolation."""
+
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab")(1),
+            Color("lab(29.571% 68.304 -112.04)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab")(0.75),
+            Color("lab(35.75% 71.43 -66.558)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab")(0.5),
+            Color("lab(41.93% 74.556 -21.079)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab")(0.25),
+            Color("lab(48.109% 77.682 24.401)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab")(0),
+            Color("lab(54.288% 80.808 69.881)")
+        )
+
+    def test_interpolate_alpha(self):
+        """Test mixing alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0 / 0.75)').interpolate('color(srgb 0 0 1 / 0.25)', space="srgb")(0.5),
+            Color('rgb(127.5 0 127.5 / 0.5)')
+        )
+
+    def test_interpolate_premultiplied_alpha(self):
+        """Test premultiplied alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0 / 0.75)').interpolate(
+                'color(srgb 0 0 1 / 0.25)', space="srgb", premultiplied=True
+            )(0.5),
+            Color('rgb(191.25 0 63.75 / 0.5)')
+        )
+
+    def test_interpolate_premultiplied_no_alpha(self):
+        """Test premultiplied alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0)').interpolate('color(srgb 0 0 1)', space="srgb", premultiplied=True)(0.5),
+            Color('color(srgb 1 0 0)').interpolate('color(srgb 0 0 1)', space="srgb")(0.5)
+        )
+
+    def test_interpolate_nan(self):
+        """Test mixing with NaN."""
+
+        c1 = Color("srgb", [NaN, 1, 1])
+        c2 = Color("srgb", [0.75, 0, 0])
+        self.assertColorEqual(c1.interpolate(c2, space="srgb")(0.5), Color("srgb", [0.75, 0.5, 0.5]))
+        c1 = Color("srgb", [0.25, 1, 1])
+        c2 = Color("srgb", [NaN, 0, 0])
+        self.assertColorEqual(c1.interpolate(c2, space="srgb")(0.5), Color("srgb", [0.25, 0.5, 0.5]))
+        c1 = Color("srgb", [NaN, 1, 1])
+        c2 = Color("srgb", [NaN, 0, 0])
+        self.assertColorEqual(c1.interpolate(c2, space="srgb")(0.5), Color("srgb", [0, 0.5, 0.5]))
+
+    def test_interpolate_adjust(self):
+        """Test mix adjust method."""
+
+        c1 = Color("color(srgb 0.25 1 1)")
+        c2 = Color("color(srgb 0.75 0 0)")
+        self.assertColorEqual(
+            c1.interpolate(c2, adjust=["green", "blue"], space="srgb")(0.5),
+            Color("srgb", [0.25, 0.5, 0.5])
+        )
+
+    def test_interpolate_hue_adjust(self):
+        """Test hue adjusting."""
+
+        c1 = Color('rebeccapurple')
+        c2 = Color('lch(85% 100 805)')
+        self.assertColorEqual(
+            c1.interpolate(c2, adjust=["hue"], hue="shorter", space="lch")(0.25),
+            Color("rgb(146.74 -3.9476 106.4)")
+        )
+        self.assertColorEqual(
+            c1.interpolate(c2, adjust=["hue"], hue="longer", space="lch")(0.25),
+            Color("rgb(-86.84 87.632 170)")
+        )
+        self.assertColorEqual(
+            c1.interpolate(c2, adjust=["hue"], hue="increasing", space="lch")(0.25),
+            Color("rgb(146.74 -3.9476 106.4)")
+        )
+        self.assertColorEqual(
+            c1.interpolate(c2, adjust=["hue"], hue="decreasing", space="lch")(0.25),
+            Color("rgb(-86.84 87.632 170)")
+        )
+        self.assertColorEqual(
+            c1.interpolate(c2, adjust=["hue"], hue="specified", space="lch")(0.25),
+            Color("rgb(112.84 63.966 -28.832)")
+        )
+
+    def test_interpolate_progress(self):
+        """Test custom progress."""
+
+        progress = lambda x: x * 3  # noqa: E731
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab", progress=progress)(1),
+            Color("lab(-19.862% 43.296 -475.88)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab", progress=progress)(0.75),
+            Color("lab(-1.3247% 52.674 -339.44)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab", progress=progress)(0.5),
+            Color("lab(17.213% 62.052 -203)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab", progress=progress)(0.25),
+            Color("lab(35.75% 71.43 -66.558)")
+        )
+        self.assertColorEqual(
+            Color('red').interpolate('blue', out_space="lab", space="lab", progress=progress)(0),
+            Color("lab(54.288% 80.808 69.881)")
+        )
+
+    def test_steps(self):
+        """Test steps."""
+
+        colors = Color('red').steps('blue', space="srgb", steps=5)
+        self.assertColorEqual(colors[4], Color("srgb", [0, 0, 1]))
+        self.assertColorEqual(colors[3], Color("srgb", [0.25, 0, 0.75]))
+        self.assertColorEqual(colors[2], Color("srgb", [0.5, 0, 0.5]))
+        self.assertColorEqual(colors[1], Color("srgb", [0.75, 0, 0.25]))
+        self.assertColorEqual(colors[0], Color("srgb", [1, 0, 0]))
+
+    def test_steps_space(self):
+        """Test steps different space."""
+
+        colors = Color('red').steps('blue', space="lab", steps=5)
+        self.assertColorEqual(colors[4], Color("rgb(-0.00065 0.00037 255)"))
+        self.assertColorEqual(colors[3], Color("rgb(144.86 -24.872 194.36)"))
+        self.assertColorEqual(colors[2], Color("rgb(193 -29.51 136.17)"))
+        self.assertColorEqual(colors[1], Color("rgb(226.9 -24.31 79.189)"))
+        self.assertColorEqual(colors[0], Color("rgb(255 0.00015 0.00002)"))
+
+    def test_steps_out_space(self):
+        """Test steps with output in different space."""
+
+        colors = Color('red').steps('blue', space="srgb", steps=5, out_space="lab")
+        self.assertColorEqual(
+            colors[4],
+            Color("color(lab 29.571 68.304 -112.04 / 1)")
+        )
+        self.assertColorEqual(
+            colors[3],
+            Color("lab(24.639% 57.343 -83.554)")
+        )
+        self.assertColorEqual(
+            colors[2],
+            Color("lab(29.562% 55.96 -36.2)")
+        )
+        self.assertColorEqual(
+            colors[1],
+            Color("lab(41.109% 66.205 23.404)")
+        )
+        self.assertColorEqual(
+            colors[0],
+            Color("lab(54.288% 80.808 69.881)")
+        )
+
+    def test_steps_alpha(self):
+        """Test mixing alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0 / 0.75)').steps('color(srgb 0 0 1 / 0.25)', space="srgb", steps=1)[0],
+            Color('rgb(127.5 0 127.5 / 0.5)')
+        )
+
+    def test_steps_premultiplied_alpha(self):
+        """Test premultiplied alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0 / 0.75)').steps(
+                'color(srgb 0 0 1 / 0.25)', space="srgb", steps=1, premultiplied=True
+            )[0],
+            Color('rgb(191.25 0 63.75 / 0.5)')
+        )
+
+    def test_steps_premultiplied_no_alpha(self):
+        """Test premultiplied alpha."""
+
+        self.assertColorEqual(
+            Color('color(srgb 1 0 0)').steps('color(srgb 0 0 1)', space="srgb", steps=1, premultiplied=True)[0],
+            Color('color(srgb 1 0 0)').steps('color(srgb 0 0 1)', space="srgb", steps=1)[0]
+        )
+
+    def test_steps_nan(self):
+        """Test steps with NaN."""
+
+        c1 = Color("srgb", [NaN, 1, 1])
+        c2 = Color("srgb", [0.75, 0, 0])
+        self.assertColorEqual(c1.steps(c2, space="srgb", steps=1)[0], Color("srgb", [0.75, 0.5, 0.5]))
+        c1 = Color("srgb", [0.25, 1, 1])
+        c2 = Color("srgb", [NaN, 0, 0])
+        self.assertColorEqual(c1.steps(c2, space="srgb", steps=1)[0], Color("srgb", [0.25, 0.5, 0.5]))
+        c1 = Color("srgb", [NaN, 1, 1])
+        c2 = Color("srgb", [NaN, 0, 0])
+        self.assertColorEqual(c1.steps(c2, space="srgb", steps=1)[0], Color("srgb", [0, 0.5, 0.5]))
+
+    def test_steps_adjust(self):
+        """Test steps with adjust method."""
+
+        c1 = Color("color(srgb 0.25 1 1)")
+        c2 = Color("color(srgb 0.75 0 0)")
+        self.assertColorEqual(
+            c1.steps(c2, adjust=["green", "blue"], space="srgb", steps=1)[0],
+            Color("srgb", [0.25, 0.5, 0.5])
+        )
+
+    def test_steps_hue_adjust(self):
+        """Test steps with hue adjusting."""
+
+        self.assertColorEqual(
+            Color('rebeccapurple').steps('lch(85% 100 805)', space="lch", steps=5, adjust=["hue"], hue="shorter")[1],
+            Color("rgb(146.74 -3.9476 106.4)")
+        )
+        self.assertColorEqual(
+            Color('rebeccapurple').steps('lch(85% 100 805)', space="lch", steps=5, adjust=["hue"], hue="longer")[1],
+            Color("rgb(-86.84 87.632 170)")
+        )
+        self.assertColorEqual(
+            Color('rebeccapurple').steps('lch(85% 100 805)', space="lch", steps=5, adjust=["hue"], hue="increasing")[1],
+            Color("rgb(146.74 -3.9476 106.4)")
+        )
+        self.assertColorEqual(
+            Color('rebeccapurple').steps('lch(85% 100 805)', space="lch", steps=5, adjust=["hue"], hue="decreasing")[1],
+            Color("rgb(-86.84 87.632 170)")
+        )
+        self.assertColorEqual(
+            Color('rebeccapurple').steps('lch(85% 100 805)', space="lch", steps=5, adjust=["hue"], hue="specified")[1],
+            Color("rgb(112.84 63.966 -28.832)")
+        )
+
+    def test_steps_progress(self):
+        """Test custom progress."""
+
+        progress = lambda x: x * 3  # noqa: E731
+        colors = Color('red').steps('blue', steps=5, out_space="lab", space="lab", progress=progress)
+        self.assertColorEqual(
+            colors[4],
+            Color("lab(-19.862% 43.296 -475.88)")
+        )
+        self.assertColorEqual(
+            colors[3],
+            Color("lab(-1.3247% 52.674 -339.44)")
+        )
+        self.assertColorEqual(
+            colors[2],
+            Color("lab(17.213% 62.052 -203)")
+        )
+        self.assertColorEqual(
+            colors[1],
+            Color("lab(35.75% 71.43 -66.558)")
+        )
+        self.assertColorEqual(
+            colors[0],
+            Color("lab(54.288% 80.808 69.881)")
+        )
+
+    def test_steps_max_delta_e(self):
+        """Test steps with a max delta e."""
+
+        colors = Color('red').steps('blue', space="srgb", max_delta_e=10)
+        for index, color in enumerate(colors, 0):
+            if not index:
+                continue
+            self.assertTrue(color.delta_e(colors[index - 1]) <= 10)
+
+        colors = Color('red').steps('blue', space="srgb", max_delta_e=3)
+        for index, color in enumerate(colors, 0):
+            if not index:
+                continue
+            self.assertTrue(color.delta_e(colors[index - 1]) <= 3)
+
+    def test_steps_max_delta_e_steps(self):
+        """Test steps with a max delta e."""
+
+        colors = Color('red').steps('blue', space="srgb", max_delta_e=10)
+        self.assertTrue(len(colors) > 5)
+        colors = Color('red').steps('blue', space="srgb", max_delta_e=10, max_steps=5)
+        self.assertTrue(len(colors) == 5)